requirements.txt
Flask==3.1.3
python-dotenv==1.0.1
openai==1.68.2
python-docx==1.1.2
reportlab==4.2.2
langgraph==0.2.60
langchain-core==0.3.45
volcengine
